// src/Models/fact_about_us.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 07 Apr 2022 08:01:06 GMT'
 */
export type FactAboutUs = IContentItem<{
  description: Elements.RichTextElement;
  title: Elements.TextElement;
  sitemap: Elements.TaxonomyElement;
  image: Elements.AssetsElement;
}>;