// src/Models/about_us.ts
import { IContentItem, Elements } from '@kentico/kontent-delivery';
import { FactAboutUs } from './fact_about_us';

/**
 * Generated by '@kentico/kontent-model-generator@4.1.0' at 'Thu, 07 Apr 2022 08:01:06 GMT'
 */
export type AboutUs = IContentItem<{
  metadataOgDescription: Elements.TextElement;
  metadataMetaTitle: Elements.TextElement;
  metadataOgTitle: Elements.TextElement;
  metadataMetaDescription: Elements.TextElement;
  metadataTwitterSite: Elements.TextElement;
  urlPattern: Elements.UrlSlugElement;
  metadataTwitterImage: Elements.AssetsElement;
  sitemap: Elements.TaxonomyElement;
  metadataTwitterCreator: Elements.TextElement;
  metadataTwitterTitle: Elements.TextElement;
  metadataTwitterDescription: Elements.TextElement;
  metadataOgImage: Elements.AssetsElement;
  facts: Elements.LinkedItemsElement<FactAboutUs>;
}>;